Some exceptions

+++ WebShop.Core/Services/Implementation/CustomerService.cs
@@ -20,12 +20,28 @@ namespace WebShop.Core.Services.Implementation
 
         public Customer CreateCustomer(Customer customer)
         {
+            if (customer.Name == null)
+            {
+                throw new ArgumentException("Name can not be empty.");
+            }
+            else if (customer.Email == null || customer.PhoneNumber.Equals(""))
+            {
+                throw new ArgumentException("A way to contact the user is a must.");
+            }
+            else if (customer.PreferredConsole != "PC" || customer.PreferredConsole != "PS4" || customer.PreferredConsole != "XBOX1")
+            {
+                throw new ArgumentException("We do not support this console yet at the store.");
+            }
             return _customRepo.CreateCustomer(customer);
         }
 
         public Customer DeleteCustomer(int idCustomer)
         {
-            return _customRepo.DeleteCustomer(idCustomer);
+            if (idCustomer > 0)
+            {
+                return _customRepo.DeleteCustomer(idCustomer);
+            }
+            throw new ArgumentException("The ID you try to delete is not available.");
         }
 
         public List<Customer> GetAllCustomers()
@@ -33,11 +49,15 @@ namespace WebShop.Core.Services.Implementation
             return _customRepo.ReadCustomer().ToList();
         }
 
-        public List<Customer> GetCustomerById(int idCustomer)
+        public Customer GetCustomerById(int idCustomer)
         {
-            var list = _customRepo.ReadCustomer();
-            var customer = list.Where(c => c.Id.Equals(idCustomer));
-            return customer.ToList();
+
+            if (idCustomer > 0)
+            {
+                var customer = _customRepo.ReadCustomerByID(idCustomer);
+                return customer;
+            }
+            throw new ArgumentException("The ID you try to get is not available.");
         }
 
         public Customer UpdateCustomer(Customer customerToUpdate)
